build-yocto-genivi: Install locale tools and fix locale
A few issues remained (or appeared in 16.04) regarding the locale.
We know since before that bitbake requires a UTF-8 enabled locale and
the system has used en_US.UTF-8 for that.  However, en_US.UTF-8 was not
available in the latest image and also the locale generation tools
were not installed to generate it.

When the requested locale is not available, apt-get also complains
during all installations.  This seems harmless but looks ugly.

Finally we _do_ need the locale once it is time to run bitbake.

To fix all this, first use a default "C" locale for the very first
apt-get to avoid complaints.  Then install locale tools so that the
en_US.UTF-8 locale can be created, and finally the locale can be set in
the environment.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>