Add single-entry script to launch a build in docker
This builds the docker image and then runs the container with the
clone-and-build script as the README suggests.  This way, a clean build
(of the master branch currently*) can be launched with a single script.

TODO: Set up the $TARGET so that different variants can be built,
and also options for choosing branch/variant etc. (that however overlaps
theh features of scripts/ci-build.sh in genivi-dev-platform repository -
so the process should probably be changed to call that script instead).

NOTE: It creates a working directory in the user's home directory

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>