Dockerfile: Remove Go Agent, use Ubuntu and simplify
This removes the support for go-agent because for the GENIVI
infrastructure we use our gocd-setup repository and to ensure
all agents behave identically, we would not enable any agent
built with this different setup anyway.

(Maybe, "easy-genivigo" project is another option for those that want to
 mess around with their own Go Setup)

This container is better suited to be a minimal setup for anyone who
only wants to build GDP or baseline in a controlled environment.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>